EventManager/requirements.txt
pendulum
python-decouple
pytz
